requirements.txt
# Web application framework
streamlit

# OpenAI API client
openai

# Environment variable management
python-dotenv

# Asynchronous HTTP requests
aiohttp

# File handling
PyPDF2
python-docx
docx2txt

# Document conversion
pypandoc

# Data manipulation and analysis
pandas

# Date and time manipulation
python-dateutil

# Progress bars
tqdm

# ZIP file handling
zipfile36